pyproject.toml
[project]
name = "ai-data-analyst"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.0.3",
    "langchain-openai>=1.0.2",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.44",
    "streamlit>=1.51.0",
]
